Replace rolimoa_extension.py with new ver (using httpx_ws)
"websocket" ライブラリ、すごく標準ですみたいな雰囲気なのに非同期処理ができなかったので、
WebSocketクライアントをhttpx_wsに変更して改善した実装に変更
